Dynamically update the applet on provider changes
Store last received data for the currently used provider,
instead of requesting the provider data again when it might
have changed. The data engine automatically pushes updated data
for the connected provider data source (since the last commit).
Two new states providerReady/NotReady have been added to the
state machine.
This way the applet can react dynamically to changes on the
installed providers, simply copy the updated provider files to
the provider installation directory. For example when a provider
becomes invalid/not ready, the applet directly shows the error
or status message. Recovery from error with a provider that was
erroneous and then became valid and ready now also works.