Forward network error messages from scripts
Add error parameters to Scripting::NetworkRequest::finished(),
throw an error from the script if a network request failed
but the script depends on the request to be finished successfully.
The exception gets detected by the engine and the error message
gets published in the data source, ie. shown in eg. the applet.