Remove deprecated network state checks in applet
The applet was using the "network" data engine to check
the network status, this did not work very well. And since
network error messages are now forwarded from the engine,
no network state checks need to be done in the applet.
Furthermore the applet cannot know if the provider plugin
needs a network connection at all (eg. GTFS does not).

Remove old error messages in the applet, just use the error
message from the data engine.