Add time zone validity test, validate in dialog
New test ServiceProviderDataTimeZoneTest in TestModel.
Add ServiceProviderDataTester::isTimeZoneValid().
Show error in ProjectSettingsDialog when the time zone is invalid.

Improve ProjectSettingsDialog::appendMessageWidgetAfter()
to show the message at the bottom of the current tab if the
widget after which the message should be shown is not visible,
ie. on another tab.

Fix time zone name "PST" to "America/Los_Angeles" for us_bart
(plus formatting done by TimetableMate).

Fix name of the main QTabWidget in timetablemateview_base.ui,
somehow that name was used twice: "tabNotes". Renamed to "tabWidget".