Require provider plugin format 1.1
Expand the meaning of the version to the whole plugin format,
not only the .pts file, eg. also the script API. Scripts for 1.0
are no longer supported and GTFS providers were not supported
at all in 1.0, therefore 1.1 is required now.

Add possibility to ServiceProviderDataReader to get an error
message without raiseError() when using ReadErroneousFiles.
This gets used in TimetableMate, error messages are now shown
also when loading the project and also for inactive projects
(warnings, info messages still only for active projects).

ProjectSettingsDialog now automatically updates to "1.1",
the user gets informed in a KMessageBox. The KLineEdit to
manually edit the plugin format version has been removed.

Update documentation, error messages