Fix loop of automatic updates with empty data
It was possible that automatic updates were done all the time
in a loop controlled by the update timer with a timeout of 0ms.
This is because the next update time gets calculated based on the
received data. Fixed by making ServiceProvider::minFetchWait() return
at least 60 (seconds). Also wrong 'last update' times were used,
they could be invalid, producing invalid next update times.

Add a Q_ASSERT to ensure that the time until the next update is
not too small.