Fix GHNS feature using a new Synchrotron provider
Available at collect.kde.org/publictransport-providers/,
very easy to setup :)

Update StopSettingsDialog, remove broken "Near Stops..."
button, the popup map can be used instead. Also remove
StopFinder class, was only used by the removed button.
Make the GHNS button a dialog button to give it a more
prominent place (and show it's text) since it will be
required to find this button to use providers,
that are not installed by default.

To ensure the local provider install dir gets monitored by
the data engine, it gets created on start with a small file
in it to prevent removal of the directory (eg. when the last
GHNS provider gets uninstalled again).

TODO: Move the other provider plugins over to the GHNS
  provider and remove them from the data engine.
  Only include some default providers.