Add data for broken providers to data sources
Only if the provider plugin .pts file cannot be read, no data can be
provided like eg. the provider name. For other errors the provider data
is now included in service provider data sources.
Also ensure that a "state" field is always available, set it to "error"
on errors and copy the "errorMessage" to "statusMessage".