Fix auto-show GHNS dialog, add "isInstalled" provider state
Add "isInstalled" field to provider state data (field "stateData")
for erroneous providers (field "error" is true). If provider data
for a nonexistant provider was requested it also shows up in the
"ServiceProviders" data source. The new field is needed to
distinguish nonexistant providers from providers with other errors.

Check in StopSettingsDialog if providers are available.
If only one provider is available, check if it is installed and
also show the GHNS dialog if it is not.