Fix rerunning completiongenerator on every CMake run
Fixed the name of an output file (datastream.html to
datastreamprototype.html). The command was rerun in every
CMake run because datastream.html was never found
(wrong filename).

Use completiongenerator target for add_custom_command
instead of specifying the output executable of that target
and manually depending on it.

Also do not put generated HTML documentation into the
source directory.