Show credit in the applet when available, remove "last update"
Some providers require special credit strings to get shown
with the timetable data. These strings can be given in
<credit> tags in the .pts files.

If available this now replaces the "data by: <url>" string.
If the credit string is too long it gets elided, the full
string gets shown in the tooltip.

Move "last update" text to the tooltip.