Fix TimetableMate completion, add "Show Docs for ..." links
Documentation generated by completiongenerator for TimetableMate
was not nicely formatted. Inline verbatim code went into it's
own output line, multiple code lines between @begincode and
@endcode went into a single output line, too many newlines at
some places, HTML tags in the documentation source were not
properly encoded.

Add "Show Documentation for ..." links in completions to show
the documentation in it's dock (like in KDevelop).

Show all function overloads for completion.

Add completiongenerator to the DEPENDS list for
add_custom_command. Did not work without it although
completiongenerator is given as COMMAND and should be known
to CMake (kde4_add_executable), at least it is found
to execute it.