Fix crash with failing additional data requests
Fix crash when an additional data request fails but additional data
requests for more timetable items are running for the same timetable
service job. Fix by moving the loop to start requests for multiple items
to the engine so that the job does nothing after starting the request in
the engine and/or by only emitting the finished() signal when the job is
completely done (not when the first job fails).

It could crashed because the finished() signals of the jobs are
connected to the deleteLater() slot of the timetable service, which is
the parent of all it's jobs. The job could emit finished() too early,
ie. while it was still running, leading to a crash.

Do not block service jobs or emit finished() from start(). Use finished
signals in the engine for all timetable service jobs, instead of using a
pointer for the result as parameter to the request methods. This can
also greatly speed up requests for additional data of multiple items.

Store pointer to the data engine in the service jobs, use new base
class.

Use "additionalDataState" and "additionalDataError" instead of
"IncludesAdditionalData", "WaitingForAdditionalData".