Fix stop names in "action buttons" when in panel
No stop names were shown in the stop tab bar in
"action button" state, when the applet was in a panel.
Stop names were elided so much, that nothing was left,
because the size of the icon in the panel was used instead
of the size of the popup.
Additional use 60 pixels as minimum text width for stop names.